requirements.txt
# requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.6.3
flask-migrate==3.1.0
flask-wtf==0.15.1
werkzeug==2.0.1
python-dotenv==0.19.0
mercadopago==2.0.8
flask-pwa==0.1.0
email-validator==1.1.3
python-dateutil==2.8.2
pillow==8.3.1
reportlab==3.6.1
gunicorn==20.1.0
flask-cors==3.0.10